+++ src/components/Code/example/index.js
@@ -15,7 +15,15 @@ var bat = 'baz';` }
     examples: [
       <Code type="block" isHighlighted={ true } language="javascript">
 { `var highlightedFoo = 'bar';
-var highLightedBat = 'baz';` }
+var highlightedBat = 'baz';` }
+      </Code>,
+    ],
+  },
+  {
+    examples: [
+      <Code type="block" hasCopyButton={ true } testSection="code">
+{ `var copyableFoo = 'bar';
+var copyableBat = 'baz';` }
       </Code>,
     ],
   },