+++ src/components/Code/tests/index.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { default as Clipboard } from 'clipboard';
 import * as testHelpers from '../../../utils/test-helpers';
 import Code from '../index';
 
@@ -20,7 +21,7 @@ describe('components/Code', () => {
 
     const componentNode = testHelpers.getNodeFromComponent(component);
 
-    expect(componentNode.tagName).toBe('PRE');
+    expect(componentNode.children[0].tagName).toBe('PRE');
   });
 
   it('should render code that is passed in', () => {
@@ -77,8 +78,7 @@ describe('components/Code', () => {
     expect(componentBlockNode.innerHTML).not.toContain('class="hljs-');
   });
 
-  it(`should not add syntax highlighting if language is provided but
-    highlighting is not requested`, () => {
+  it('should not add syntax highlighting if language is provided but highlighting is not requested', () => {
     let code = 'var foo;';
 
     const componentInline = testHelpers.renderIntoDocument(
@@ -96,6 +96,102 @@ describe('components/Code', () => {
     expect(componentBlockNode.innerHTML).not.toContain('class="hljs-');
   });
 
+  it('should not add a copy button by default', () => {
+    let code = 'var foo;';
+
+    const component = testHelpers.renderIntoDocument(
+      <Code
+        type="block"
+        testSection="code">
+        { code }
+      </Code>
+    );
+
+    const copyButton = testHelpers.getTestSectionFromComponent(component, 'code-copy-button');
+
+    expect(copyButton).toBeNull();
+  });
+
+  describe('with copy functionality', () => {
+    it('should add a copy button to block code', () => {
+      let code = 'var foo;';
+
+      const component = testHelpers.renderIntoDocument(
+        <Code
+          type="block"
+          hasCopyButton={ true }
+          testSection="code">
+          { code }
+        </Code>
+      );
+
+      const copyButton = testHelpers.getTestSectionFromComponent(component, 'code-copy-button');
+
+      expect(copyButton).not.toBeNull();
+      expect(copyButton.type).toBe('button');
+    });
+
+    it('should call clipboard function when button is clicked on', () => {
+      let code = 'var foo;';
+
+      const component = testHelpers.renderIntoDocument(
+        <Code
+          type="block"
+          hasCopyButton={ true }
+          testSection="code">
+          { code }
+        </Code>
+      );
+
+      spyOn(Clipboard.prototype, 'onClick').and.stub();
+      spyOn(Clipboard.prototype, 'resolveOptions').and.stub();
+
+      const copyButton = testHelpers.getTestSectionFromComponent(component, 'code-copy-button');
+      testHelpers.simulate.click(copyButton);
+
+      expect(Clipboard.prototype.resolveOptions.calls.argsFor(0)[0].text()).toEqual(code);
+      expect(Clipboard.prototype.onClick.calls.count()).toEqual(1);
+    });
+
+    it('should destroy clipboard after button is clicked on', () => {
+      let code = 'var foo;';
+
+      const component = testHelpers.renderIntoDocument(
+        <Code
+          type="block"
+          hasCopyButton={ true }
+          testSection="code">
+          { code }
+        </Code>
+      );
+
+      spyOn(Clipboard.prototype, 'destroy').and.stub();
+
+      const copyButton = testHelpers.getTestSectionFromComponent(component, 'code-copy-button');
+      testHelpers.simulate.click(copyButton);
+
+      expect(Clipboard.prototype.destroy.calls.count()).toEqual(1);
+    });
+
+    it('should not add a copy button to inline code', () => {
+      let code = 'var foo;';
+
+      const component = testHelpers.renderIntoDocument(
+        <Code
+          type="inline"
+          hasCopyButton={ true }
+          testSection="code">
+          { code }
+        </Code>
+      );
+
+      const copyButton = testHelpers.getTestSectionFromComponent(component, 'code-copy-button');
+
+      expect(copyButton).toBeNull();
+    });
+  });
+
+
   it('should have a properly set test section', () => {
     const componentInline = testHelpers.renderIntoDocument(
       <Code testSection="foo-inline" type="inline">Hello!</Code>
@@ -109,6 +205,6 @@ describe('components/Code', () => {
     const componentBlockNode = testHelpers.getNodeFromComponent(componentBlock);
 
     testHelpers.expectTestSectionToExist(componentInlineNode, 'foo-inline');
-    testHelpers.expectTestSectionToExist(componentBlockNode, 'foo-block');
+    testHelpers.expectTestSectionToExist(componentBlockNode.children[0], 'foo-block');
   });
 });