+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 ### Added
 - [Feature] Add a `tiny` sized button. (#454)
 - [Patch] Add more examples to the React documentation. (#460)
+- [Feature] Add optional "Copy" button to Code component. (#453)
 
 ### Fixed
 - [Patch] Fix tiny README mistakes. (#451)